[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17716 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,9 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int a = scanner.nextInt();
-        int b = scanner.nextInt();
-
+        long a = scanner.nextLong();
+        long b = scanner.nextLong();
+        
         System.out.println((a + b) * (a - b));
+        
     }
 }
